+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/rule/GeckoSessionTestRule.java
@@ -2955,6 +2955,14 @@ public class GeckoSessionTestRule implements TestRule {
 
 
 
+  public @Nullable JSONObject getIPPProxyInfo() {
+    return (JSONObject) webExtensionApiCall("GetIPPProxyInfo", null);
+  }
+
+  
+
+
+