+++ mobile/shared/modules/geckoview/GeckoViewIPProtection.sys.mjs
@@ -185,7 +185,11 @@ export const GeckoViewIPProtection = {
         break;
       }
       case "GeckoView:IPProtection:Activate": {
-        lazy.IPPProxyManager.start()
+        lazy.IPPProxyManager.start(
+          aData?.userAction ?? true,
+          aData?.inPrivateBrowsing ?? false,
+          aData?.country
+        )
           .then(({ started, error } = {}) => {
             if (started) {
               aCallback.onSuccess();